[Fix] DaangnMarket#33 - 매너온도 모델 Int -> Float 변경

+++ DaangnMarket/DaangnMarket/Presentation/Product/Model/ProductModel.swift
@@ -14,7 +14,7 @@ struct ProductModel {
     var profileImage: UIImage
     var nickname: String
     var address: String
-    var mannerTemp: Int
+    var mannerTemp: Float
     var title: String
     var category: String
     var createdAt: String
